scripts/requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
hdbscan==0.8.41
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
joblib==1.5.3
llvmlite==0.46.0
numba==0.63.1
numpy==2.3.5
openai==2.17.0
openpyxl==3.1.5
pandas==3.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pynndescent==0.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
scikit-learn==1.8.0
scipy==1.17.0
six==1.17.0
sniffio==1.3.1
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
umap-learn==0.5.11
